+++ wasm/src/sub_account.rs
@@ -18,7 +18,10 @@ use crate::errors::WasmError;
 /// const sub = deriveSubAccountAddress(master, 1);
 /// const account = await client.accountInfo(sub);
 /// await client.sendCallMessage(User.placeOrders(marketId, orders, false, 1));
+//
+// `index` is `u32` (not `u8`) so a JS number outside `0..=255` is validated by
+// the range check rather than silently truncated mod 256 at the JS→wasm boundary.
 #[wasm_bindgen(js_name = deriveSubAccountAddress)]
-pub fn derive_sub_account_address(master: &str, index: u8) -> Result<String, WasmError> {
+pub fn derive_sub_account_address(master: &str, index: u32) -> Result<String, WasmError> {
     bullet_rust_sdk::derive_sub_account_address(master, index).map_err(WasmError::from)
 }
